requirements.txt
black==25.1.0
click==8.1.8
colorama==0.4.6
dotenv==0.9.9
greenlet==3.1.1
mimesis==18.0.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
psycopg==3.2.6
psycopg-binary==3.2.6
python-dotenv==1.1.0
PyYAML==6.0.2
SQLAlchemy==2.0.40
typing_extensions==4.13.0
tzdata==2025.2
